feat: agregar pruebas y fixtures para la gestión de la base de datos SQLite
