+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 pythonpath = src
-testpaths = test
+testpaths = tests
