+++ tests/integration/test_database.py
@@ -6,21 +6,6 @@ import config.rutas as ruta
 from infraestructura.persistencia.database_manager import SQLiteManager
 
 
-@pytest.fixture
-def db_conexion():
-    conexion = sqlite3.connect(":memory:")
-    cursor = conexion.cursor()
-    cursor.execute("PRAGMA foreign_keys = ON;")
-
-    with open(ruta.SCHEMA_SQL, "r") as schema:
-        cursor.executescript(schema.read())
-
-    with open(ruta.VISTA_SQL, "r") as schema:
-        cursor.executescript(schema.read())
-
-    return conexion
-
-
 def test_schema_sql_ejecuta_sin_errores():
     """
     Testea que se ejecute el schema en una DB vacia sin generar excepciones
@@ -43,9 +28,6 @@ def test_view_sql_ejecuta_sin_errores():
     with open(ruta.VISTA_SQL, "r", encoding="utf-8") as f:
         conexion.executescript(f.read())
     conexion.close()
-
-
-# test/test_database.py
 
 
 def test_schema_sql_ejecuta_sin_errores():
@@ -76,7 +58,7 @@ def test_seed_sql_ejecuta_sin_errores():
     manager.close_connection()
 
 
-def test_database_schema(db_conexion):
+def test_database_schema(db_conexion_sin_seed):
     """
     Test destinado a verificar que se creen las tablas especificadas en en schenma
     """
@@ -96,7 +78,7 @@ def test_database_schema(db_conexion):
         "categoria",
     }
 
-    cursor = db_conexion.cursor()
+    cursor = db_conexion_sin_seed.cursor()
     cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
 
     tablas_existentes = {row[0] for row in cursor.fetchall()}
@@ -105,7 +87,7 @@ def test_database_schema(db_conexion):
     assert not faltan, f"Error: No se encontraron las tablas: {faltan}"
 
 
-def test_database_schema_view(db_conexion):
+def test_database_schema_view(db_conexion_sin_seed):
     """
     Test destinado a verificar que se generen las vistas especificadas en el archivo sql
     """
@@ -117,7 +99,7 @@ def test_database_schema_view(db_conexion):
         "v_partidos_resumen",
     }
 
-    cursor = db_conexion.cursor()
+    cursor = db_conexion_sin_seed.cursor()
     cursor.execute("SELECT name FROM sqlite_master WHERE type='view';")
 
     vistas_existentes = {row[0] for row in cursor.fetchall()}
@@ -126,10 +108,10 @@ def test_database_schema_view(db_conexion):
     assert not faltan, f"Error: No se encontraron las vistas: {faltan}"
 
 
-def test_referential_integrity(db_conexion):
+def test_referential_integrity(db_conexion_sin_seed):
     # Intentar insertar un registro con una categoría que NO existe (id: 999)
     # Debe lanzar un sqlite3.IntegrityError
-    db_cursor = db_conexion.cursor()
+    db_cursor = db_conexion_sin_seed.cursor()
     with pytest.raises(sqlite3.IntegrityError):
         db_cursor.execute(
             "INSERT INTO listaBuenaFe (fechaPresentacion, idInscripcion) VALUES (?, ?)",
@@ -137,12 +119,12 @@ def test_referential_integrity(db_conexion):
         )
 
 
-def test_check_constraints(db_conexion):
+def test_check_constraints(db_conexion_sin_seed):
     """
     Test destinado a verificar que se cumplan los CHECK definidos en la DB
     """
 
-    db_cursor = db_conexion.cursor()
+    db_cursor = db_conexion_sin_seed.cursor()
     db_cursor.execute("PRAGMA foreign_keys = OFF;")
 
     # 1. Probar puntos negativos
@@ -282,8 +264,8 @@ def test_division_by_zero_devuelve_cero_en_vistas():
     assert porcentaje_t1 == 0.0
 
 
-def test_business_constraints(db_conexion):
-    db_cursor = db_conexion.cursor()
+def test_business_constraints(db_conexion_sin_seed):
+    db_cursor = db_conexion_sin_seed.cursor()
 
     # Pre-requisitos: Clubes y Competencia
     db_cursor.execute("INSERT INTO club (nombre) VALUES ('Atenas')")
@@ -329,8 +311,8 @@ def test_business_constraints(db_conexion):
         )
 
 
-def test_foreign_key_behavior(db_conexion):
-    cursor = db_conexion.cursor()
+def test_foreign_key_behavior(db_conexion_sin_seed):
+    cursor = db_conexion_sin_seed.cursor()
 
     # Setup datos iniciales
     cursor.execute("INSERT INTO club (nombre) VALUES ('Atenas')")
@@ -367,7 +349,7 @@ def test_foreign_key_behavior(db_conexion):
         cursor.execute("DELETE FROM club WHERE idClub = 1")
 
 
-def test_view_semantics_and_cardinality(db_conexion):
+def test_view_semantics_and_cardinality():
     # Setup: Inicializar con datos semilla
     manager = SQLiteManager(":memory:", ruta.SCHEMA_SQL, ruta.VISTA_SQL, ruta.SEED_SQL)
     conexion = manager.connect()
